moved color stuff into its own folder.
also, other fixes in calls to verbose level, and re-working a bit of the Makemodule files.
