+++ src/body/data_type.cpp
@@ -3849,128 +3849,6 @@ int compare_integers_increasing(const void * left_in, const void * right_in)
 
 
 
-namespace color {
-	int color_to_int(const std::string c)
-	{
-		
-		if(c.compare("k") == 0){
-			return 30;
-		}
-		else if(c.compare("r") == 0){
-			return 31;
-		}
-		else if(c.compare("g") == 0){
-			return 32;
-		}
-		else if(c.compare("y") == 0){
-			return 33;
-		}
-		else if (c.compare("b") == 0) {
-			return 34;
-		}
-		else if(c.compare("m") == 0){
-			return 35;
-		}
-		else if(c.compare("c") == 0){
-			return 36;
-		}
-		else if(c.compare("l") == 0){
-			return 37;
-		}
-		
-		return 30;
-		
-	}
-	
-	
-	std::string bold(std::string new_color)
-	{
-		std::stringstream ss;
-		ss << "\033[1;" << color_to_int(new_color) << "m" ;     //;
-		return ss.str();
-	}
-	
-	std::string dark(std::string new_color)
-	{
-		std::stringstream ss;
-		ss << "\033[2;" << color_to_int(new_color) << "m" ;     //;
-		return ss.str();
-	}
-	
-	
-	std::string underline(std::string new_color)
-	{
-		std::stringstream ss;
-		ss << "\033[4;" << color_to_int(new_color) << "m" ;     //;
-		return ss.str();
-	}
-	
-	
-	std::string background(std::string new_color)
-	{
-		std::stringstream ss;
-		ss << "\033[7;" << color_to_int(new_color) << "m" ;     //;
-		return ss.str();
-	}
-	
-	
-	std::string strike(std::string new_color)
-	{
-		std::stringstream ss;
-		ss << "\033[9;" << color_to_int(new_color) << "m" ;     //;
-		return ss.str();
-	}
-	
-	std::string console_default(){
-		return "\033[0m";
-	}
-	
-	std::string black(){
-		return "\033[0;30m";
-	}
-	
-	std::string red(){
-		return "\033[0;31m";
-	}
-	
-	std::string green(){
-		return "\033[0;32m";
-	}
-	
-	
-	std::string brown(){
-		return "\033[0;33m";
-	}
-	
-	std::string blue(){
-		return "\033[0;34m";
-	}
-	
-	std::string magenta(){
-		return "\033[0;35m";
-	}
-	
-	std::string cyan(){
-		return "\033[0;36m";
-	}
-	
-	std::string gray(){
-		return "\033[0;37m";
-	}
-	
-	
-	//black - 30
-	//red - 31
-	//green - 32
-	//brown - 33
-	//blue - 34
-	//magenta - 35
-	//cyan - 36
-	//lightgray - 37
-}
-
-
-
